+++ contracts/src/validation/hooks/DelayHook.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 pragma solidity 0.8.25;
 
-import {Scheduler} from 'src/execution/Scheduler.sol';
+import {Scheduler} from '~/execution/Scheduler.sol';
 
 struct DelayConfig {
   uint32 delay /* seconds */;
@@ -15,7 +15,7 @@ library DelayHook {
     DelayConfig memory config = abi.decode(configData, (DelayConfig));
     if (config.delay == 0) return true;
 
-    Scheduler.schedule(proposal, uint32(block.timestamp) + config.delay);
+    Scheduler.schedule(proposal, uint64(block.timestamp) + config.delay);
     return false;
   }
 }
